Put URNAssetMapper in a dedicated file
